use importlib_resources to load json schema
