+++ aiida_crystal17/validation/schemas/fitting.base.schema.json
