+++ aiida_crystal17/validation/schemas/fitting.lj.schema.json
