+++ aiida_crystal17/validation/schemas/fitting_settings.schema.json
