+++ aiida_crystal17/validation/schemas/gulp_optimize.schema.json
