+++ aiida_crystal17/validation/schemas/potential.base.schema.json
