+++ aiida_crystal17/validation/schemas/potential.lj.schema.json
