+++ aiida_crystal17/validation/schemas/potential.reaxff.schema.json
