+++ aiida_crystal17/validation/schemas/structure.schema.json
