+++ aiida_crystal17/validation/schemas/symmetry.schema.json
